Nyhet. Egen hook for å sette stillingsfeltene.
